phase-05-ai-agent/requirements.txt
groq
napalm
netmiko
pyyaml
jinja2
colorama
python-dotenv
